Do not close the window automatically on disconnect if an error has been logged.
This is still not used, but it is intended to be used in cases where for example
codec negotiation fails and the channel closes because there are no more streams.
In this case an error should be reported to the user so that he has a hint and
the window should not close, to allow the user see what is wrong.

svn path=/trunk/playground/network/kcall/; revision=1003698