Import the new libtelepathy-kde-call library.
This library wraps all the gstreamer and farsight bits and hides them behind
a minimalistic API that is meant to be used by GUIs to implement audio/video
streaming. The design was chosen to allow in the future extra capabilities
such as stream capturing, detecting who is speaking, etc...

It is mostly based on old code, but it has been ported to the new
QtGStreamer api and some things have changed. Still untested, due to the
lack of a GUI.

svn path=/trunk/playground/network/kcall/; revision=1179470