Implement a new DeviceElementFactory class for creating the source/sink elements.
This new factory is capable of using phonon's settings when it is running
under a KDE session. However, it will only report alsa/oss devices that
are detected by phononserver (which should be fine for use in a kde environment
though). Of course, it uses pulseaudio if available. Under other DEs, it will
use the gconf elements. This is probably a better choice for GNOME or XFCE.

svn path=/trunk/playground/network/kcall/; revision=1181918