Do not use Tp::ContactPtr as the key in the participants map for now.
Unfortunately tp-qt4 and tp-farsight have independent handlers for the
relevant dbus signals and it is possible that tp-qt4 receives the stream
removal signal before tp-farsight does. This effectively removes the remote
contact from the group before the {audio,video}SrcPadRemoved() signals are
fired and then it asserts in findRemoteContact() as there is no remote contact
anymore from tp-qt4's pov.