Create the participant inside the onRequestResource slot.
This is because the connection might fail and we may never get the src-pad-added
signal, however we are going to get the free-resource signal, which will try to
remove the inexistent participant. To fix this, we can create the participant
earlier, even if the connection is going to fail and we are going to remove it again.