Fix the toggleDtmf function
With the current logic it would exit the fullscreen and then enter it
again.

Quoting Ekaitz "If you show the dtmf and you are in fullscreen it is
covered by the main window and you can't see it. This returns the window
to the normal state before the dialpad is shown. (Or that was the idea
when I did that...)"

Reviewed-by: Diane Trout