Prevent crashes from changes in how Bins are removed from the pipelines.
GStreamer reported warnings when trying to remove the sending bins
from the pipelines as they seem to be already unlinked when the
SinkManager::onPadUnlinked is called.

I discovered this by getting the source Element srcPad->parentElement()
and checking to see if that has a parent. After the bin is added to the
pipline it does have a parent, but by onPadRemoved it had been removed
from the pipeline.

I only tested audio calls which showed me that channelHandler()->pipeline()
could be null. I decided that that probably could happen to the video
path as well so added the guard condition there as well.